pyproject.toml
[project]
name = "actor-manager-backend"
version = "0.1.0"
description = "Backend runtime and tooling for the Actor Manager project"
requires-python = ">=3.12,<3.14"
dependencies = [
    "aiopg>=1.4.0",
    "alembic>=1.16.4",
    "asyncpg>=0.30.0",
    "pillow>=11.0.0",
    "fastapi>=0.116.1",
    "langchain>=0.3.27",
    "langchain-core>=0.3.79",
    "langchain-openai>=0.3.35",
    "minio>=7.2.18",
    "peewee-async>=1.2.1",
    "psycopg2-binary>=2.9.10",
    "pydantic>=2.11.7",
    "pyyaml>=6.0.1",
    "python-multipart>=0.0.20",
    "sqlalchemy>=2.0.42",
    "uvicorn[standard]>=0.35.0",
]

[dependency-groups]
dev = [
    "pytest>=8.4.1",
    "pytest-asyncio>=1.1.0",
]

[build-system]
requires = ["setuptools>=69", "wheel"]
build-backend = "setuptools.build_meta"

[tool.setuptools.packages.find]
where = ["."]
include = ["backend*"]
exclude = ["backend.tests*", "frontend*"]

[tool.pytest.ini_options]
pythonpath = ["."]
testpaths = ["backend/tests"]
